+++ src/njs/src/njsOracle.h
@@ -77,6 +77,7 @@ using namespace v8;
                                       (NJS_NODE_ORACLEDB_MINOR * 100) +   \
                                       (NJS_NODE_ORACLEDB_PATCH) )
 
+
 class Oracledb: public ObjectWrap
 {
  public:
@@ -86,20 +87,22 @@ class Oracledb: public ObjectWrap
    // Oracledb class
    static void Init(Handle<Object> target);
 
-   dpi::Env*    getDpiEnv () const { return dpienv_; }
-   bool         getAutoCommit () const  { return autoCommit_; }
-   unsigned int getOutFormat () const { return outFormat_; }
-   unsigned int getMaxRows ()  const  { return maxRows_; }
-   unsigned int getStmtCacheSize ()  const  { return stmtCacheSize_; }
-   unsigned int getPoolMin () const  { return poolMin_; }
-   unsigned int getPoolMax () const  { return poolMax_; }
-   unsigned int getPoolIncrement () const  { return poolIncrement_; }
-   unsigned int getPoolTimeout () const  { return poolTimeout_; }
-   unsigned int getPrefetchRows () const  { return prefetchRows_; }
+   dpi::Env*          getDpiEnv () const          { return dpienv_; }
+   bool               getAutoCommit () const      { return autoCommit_; }
+   unsigned int       getOutFormat () const       { return outFormat_; }
+   unsigned int       getMaxRows ()  const        { return maxRows_; }
+   unsigned int       getStmtCacheSize ()  const  { return stmtCacheSize_; }
+   unsigned int       getPoolMin () const         { return poolMin_; }
+   unsigned int       getPoolMax () const         { return poolMax_; }
+   unsigned int       getPoolIncrement () const   { return poolIncrement_; }
+   unsigned int       getPoolTimeout () const     { return poolTimeout_; }
+   unsigned int       getPrefetchRows () const    { return prefetchRows_; }
    const std::string& getConnectionClass () const { return connClass_; }
-   const DataType *getFetchAsStringTypes () const;
-   unsigned int getFetchAsStringTypesCount () const
-     {  return fetchAsStringTypesCount_ ;   }
+
+   const DataType*    getFetchAsStringTypes () const;
+
+   unsigned int       getFetchAsStringTypesCount () const
+   {  return fetchAsStringTypesCount_ ;   }
 
 private:
    // Define Oracledb Constructor
@@ -132,6 +135,7 @@ private:
    static NAN_PROPERTY_GETTER(GetPrefetchRows);
    static NAN_PROPERTY_GETTER(GetFetchAsString);
    static NAN_PROPERTY_GETTER(GetLobPrefetchSize);
+   static NAN_PROPERTY_GETTER(GetOracleClientVersion);
 
    // Define Setter Accessors to Properties
    static NAN_SETTER(SetPoolMin);
@@ -148,6 +152,8 @@ private:
    static NAN_SETTER(SetPrefetchRows);
    static NAN_SETTER(SetFetchAsString);
    static NAN_SETTER(SetLobPrefetchSize);
+   static NAN_SETTER(SetOracleClientVersion);
+
 
    Oracledb();
    ~Oracledb();
@@ -170,6 +176,7 @@ private:
    DataType     *fetchAsStringTypes_;   
    unsigned int fetchAsStringTypesCount_;
    unsigned int lobPrefetchSize_;
+   unsigned int oraClientVer_;
 };
 
 /**
@@ -200,6 +207,7 @@ typedef struct connectionBaton
   dpi::Env*   dpienv;
   dpi::Conn*  dpiconn;
   dpi::SPool* dpipool;
+
 
   Oracledb *oracledb;